+++ src/Nooskewl_Engine/gui.cpp
@@ -138,7 +138,7 @@ void Title_GUI::callback(void *data)
 {
 	if (loading) {
 		loading = false;
-		loaded = (Save_Load_GUI::Save_Load)((int)data) == Save_Load_GUI::LOAD;
+		loaded = (Save_Load_GUI::Save_Load)(int64_t)data == Save_Load_GUI::LOAD;
 	}
 }
